WRL/requirements-fix-version.txt
pytorch_lightning==1.3.3
# torch==1.7.1
segmentation_models_pytorch==0.1.3
imgaug==0.4.0
omegaconf==2.0.5
pandas==1.3.0
# scipy==1.3.0
libtiff==0.4.2
Pillow==8.2.0
tqdm==4.50.2
mlflow==1.13.1
protobuf==3.20.0
torchmetrics==0.3.2
# git+https://github.com/PyTorchLightning/metrics.git